+++ Capitulo 10/Ex-04.c
@@ -6,8 +6,8 @@ três colunas. Imprima o endereço de cada posição dessa matriz. */
 int main() {
     float matriz[3][3];
 
-    for (size_t i = 0; i < 3; i++) {
-        for (size_t j = 0; j < 3; j++) {
+    for (int i = 0; i < 3; i++) {
+        for (int j = 0; j < 3; j++) {
             printf("Endereco [%d][%d]: 0x%X\n", i, j, &matriz[i][j]);
         }
     }
