+++ Capitulo 10/Ex-06.c
@@ -9,7 +9,7 @@ int main() {
 
     printf("Informe os valores para o array de 5 posicoes\n\n");
 
-    for (size_t i = 0; i < 5; i++) {
+    for (int i = 0; i < 5; i++) {
         printf("[%d]: ", i);
         scanf("%d", (array+i));
 
@@ -19,7 +19,7 @@ int main() {
 
     printf("\nEndereco das posicoes contendo valores pares:\n\n");
 
-    for (size_t i = 0; i < 5; i++) {
+    for (int i = 0; i < 5; i++) {
         if (*(array+i) % 2 == 0) {
             printf("0x%X - [%d]: %d\n", (array+i), i, *(array+i));
         }
